bug : run bug

+++ src/main/java/com/example/jjangushrine/domain/cart/dto/response/CartItemCreateRes.java
@@ -6,10 +6,10 @@ import lombok.Getter;
 public class CartItemCreateRes {
     private final Long productId;
     private final String productName;
-    private final Long quantity;
+    private final int quantity;
     private final int price;
 
-    public CartItemCreateRes(Long productId, String productName, Long quantity, int price) {
+    public CartItemCreateRes(Long productId, String productName, int quantity, int price) {
         this.productId = productId;
         this.productName = productName;
         this.quantity = quantity;
